data/synthetic/pyproject.toml
[project]
name = "synthetic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ctgan>=0.11.1",
    "dotenv>=0.9.9",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "sdv>=1.30.0",
    "supabase>=2.27.0",
]
